Complete pipe syntax in robot files
The conversion to the pipe syntax did not include converting the test
and keyword title lines. This works, but does not follow the conventions
in the robot docs and visually breaks the tables.

Prepend the test and keyword title lines with the pipe character, and
replace blank lines in a given table with a single pipe character so
the robot files have consistent syntax.