docs: Remove cookiecutter from quick start guide
We no longer use cookiecutter to make the scenarios, just use the ones
in the tree.

Move the "advanced" customization features out of the quick start guide
to reduce clutter, and move them to a new paged called "customization".

Update the example molecule yaml, required by the changes in the
molecule-robotframework plugin.