Add tox.ini configuration
Add a tox.ini configuration to run the tests and library with different
versions of python and robotframework.

Add the test environments:

- Run robot with recent versions of python

- Run library unit tests with pytest, including a hello world pytest to
  git started.

- Run library code lint check with pylint

- Run sphynx to generate html docs from rst files

To use this, install tox (4.0 or later) using pipx, your package
manager, or in a virtualenv.  Run `tox list` to show the test envs.  Run
`tox -e <name>` to run a specific test env.